Fix source file order dependent trait issue
Bug fixes:
- Fix issue where classes could intermittently fail to correctly implement traits depending on exact order source files were visited in (close #526)

Technical improvements:
- The underlying cause of #526 appears to be with Mono's ILAsm: given a particular IL assember input that triggers the problem, other IL files where the content is identical except for the order of `.class` definitions may compile successfully, and if the problem IL file is assembled with the .NET Core IL assember, the issue does not occur. This commit fixes the issue by switching over to using the .NET Core IL assembler instead - provided that NuGet is available, the installer will fetch the `runtime.linux-x64.Microsoft.NETCore.ILAsm` and install along with the compiler.
- Remove code which sorted types by inheritance dependency order as it did not fix #526 in all cases on Mono's ILAsm and it's not required for the .NET Core ILAsm

+++ build/build.sh
@@ -34,5 +34,9 @@ fi
 echo $SOURCE_FILES | xargs mono $GHUL $DEBUG_OPTION -p $LIB -o ghul-new.exe
 
 mv ghul-new.exe ghul.exe
-mv ghul-new.exe.mdb ghul.exe.mdb
+
+if [ -f ghul-new.exe.mdb ] ; then
+    mv ghul-new.exe.mdb ghul.exe.mdb
+fi
+
 mv ghul-new.runtimeconfig.json ghul.runtimeconfig.json
